# go/model_new_user.go
/*
 * Payments API Template
 *
 * A sample openAPI specification that uses a simple user group system as an example to generate a skelton API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type NewUser struct {

	// 
	Name string `json:"name"`

	// 
	Tag string `json:"tag,omitempty"`
}