# go/model_user.go
/*
 * Payments API Template
 *
 * A sample openAPI specification that uses a simple user group system as an example to generate a skelton API
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type User struct {

	// 
	Name string `json:"name"`

	// 
	Id int64 `json:"id"`

	// 
	Tag string `json:"tag,omitempty"`
}